ORION-3605 Support `parser` in `kafka_source` `doublecloud_transfer_endpoint` resource (#18)
Add the `parser` field to `doublecloud_transfer_endpoint` `kafka_source`
resource. The new field is added as a block (and a set of underlying
blocks).

`kafka_source` is modified and refactored in order for tests to complete
properly. Tests are refactored to test target and source endpoint
resources in different functions.